+++ lib/local/FaceAnalyser/src/Face_utils.cpp
@@ -50,12 +50,12 @@ namespace FaceAnalysis
 {
 
 	// Pick only the more stable/rigid points under changes of expression
-	void extract_rigid_points(cv::Mat_<double>& source_points, cv::Mat_<double>& destination_points)
+	void extract_rigid_points(cv::Mat_<float>& source_points, cv::Mat_<float>& destination_points)
 	{
 		if(source_points.rows == 68)
 		{
-			cv::Mat_<double> tmp_source = source_points.clone();
-			source_points = cv::Mat_<double>();
+			cv::Mat_<float> tmp_source = source_points.clone();
+			source_points = cv::Mat_<float>();
 
 			// Push back the rigid points (some face outline, eyes, and nose)
 			source_points.push_back(tmp_source.row(1));
@@ -83,8 +83,8 @@ namespace FaceAnalysis
 			source_points.push_back(tmp_source.row(46));
 			source_points.push_back(tmp_source.row(47));
 
-			cv::Mat_<double> tmp_dest = destination_points.clone();
-			destination_points = cv::Mat_<double>();
+			cv::Mat_<float> tmp_dest = destination_points.clone();
+			destination_points = cv::Mat_<float>();
 
 			// Push back the rigid points
 			destination_points.push_back(tmp_dest.row(1));
@@ -115,16 +115,16 @@ namespace FaceAnalysis
 	}
 
 	// Aligning a face to a common reference frame
-	void AlignFace(cv::Mat& aligned_face, const cv::Mat& frame, const cv::Mat_<double>& detected_landmarks, cv::Vec6d params_global, const PDM& pdm, bool rigid, double sim_scale, int out_width, int out_height)
+	void AlignFace(cv::Mat& aligned_face, const cv::Mat& frame, const cv::Mat_<float>& detected_landmarks, cv::Vec6d params_global, const PDM& pdm, bool rigid, float sim_scale, int out_width, int out_height)
 	{
 		// Will warp to scaled mean shape
-		cv::Mat_<double> similarity_normalised_shape = pdm.mean_shape * sim_scale;
+		cv::Mat_<float> similarity_normalised_shape = pdm.mean_shape * sim_scale;
 	
 		// Discard the z component
 		similarity_normalised_shape = similarity_normalised_shape(cv::Rect(0, 0, 1, 2*similarity_normalised_shape.rows/3)).clone();
 
-		cv::Mat_<double> source_landmarks = detected_landmarks.reshape(1, 2).t();
-		cv::Mat_<double> destination_landmarks = similarity_normalised_shape.reshape(1, 2).t();
+		cv::Mat_<float> source_landmarks = detected_landmarks.reshape(1, 2).t();
+		cv::Mat_<float> destination_landmarks = similarity_normalised_shape.reshape(1, 2).t();
 
 		// Aligning only the more rigid points
 		if(rigid)
@@ -132,6 +132,7 @@ namespace FaceAnalysis
 			extract_rigid_points(source_landmarks, destination_landmarks);
 		}
 
+		// TODO rem the doubles here
 		cv::Matx22d scale_rot_matrix = AlignShapesWithScale(source_landmarks, destination_landmarks);
 		cv::Matx23d warp_matrix;
 
@@ -154,16 +155,16 @@ namespace FaceAnalysis
 	}
 
 	// Aligning a face to a common reference frame
-	void AlignFaceMask(cv::Mat& aligned_face, const cv::Mat& frame, const cv::Mat_<double>& detected_landmarks, cv::Vec6d params_global, const PDM& pdm, const cv::Mat_<int>& triangulation, bool rigid, double sim_scale, int out_width, int out_height)
+	void AlignFaceMask(cv::Mat& aligned_face, const cv::Mat& frame, const cv::Mat_<float>& detected_landmarks, cv::Vec6f params_global, const PDM& pdm, const cv::Mat_<int>& triangulation, bool rigid, float sim_scale, int out_width, int out_height)
 	{
 		// Will warp to scaled mean shape
-		cv::Mat_<double> similarity_normalised_shape = pdm.mean_shape * sim_scale;
+		cv::Mat_<float> similarity_normalised_shape = pdm.mean_shape * sim_scale;
 	
 		// Discard the z component
 		similarity_normalised_shape = similarity_normalised_shape(cv::Rect(0, 0, 1, 2*similarity_normalised_shape.rows/3)).clone();
 
-		cv::Mat_<double> source_landmarks = detected_landmarks.reshape(1, 2).t();
-		cv::Mat_<double> destination_landmarks = similarity_normalised_shape.reshape(1, 2).t();
+		cv::Mat_<float> source_landmarks = detected_landmarks.reshape(1, 2).t();
+		cv::Mat_<float> destination_landmarks = similarity_normalised_shape.reshape(1, 2).t();
 
 		// Aligning only the more rigid points
 		if(rigid)
@@ -171,18 +172,18 @@ namespace FaceAnalysis
 			extract_rigid_points(source_landmarks, destination_landmarks);
 		}
 
-		cv::Matx22d scale_rot_matrix = AlignShapesWithScale(source_landmarks, destination_landmarks);
-		cv::Matx23d warp_matrix;
+		cv::Matx22f scale_rot_matrix = AlignShapesWithScale(source_landmarks, destination_landmarks);
+		cv::Matx23f warp_matrix;
 
 		warp_matrix(0,0) = scale_rot_matrix(0,0);
 		warp_matrix(0,1) = scale_rot_matrix(0,1);
 		warp_matrix(1,0) = scale_rot_matrix(1,0);
 		warp_matrix(1,1) = scale_rot_matrix(1,1);
 
-		double tx = params_global[4];
-		double ty = params_global[5];
+		float tx = params_global[4];
+		float ty = params_global[5];
 
-		cv::Vec2d T(tx, ty);
+		cv::Vec2f T(tx, ty);
 		T = scale_rot_matrix * T;
 
 		// Make sure centering is correct
@@ -192,7 +193,7 @@ namespace FaceAnalysis
 		cv::warpAffine(frame, aligned_face, warp_matrix, cv::Size(out_width, out_height), cv::INTER_LINEAR);
 
 		// Move the destination landmarks there as well
-		cv::Matx22d warp_matrix_2d(warp_matrix(0,0), warp_matrix(0,1), warp_matrix(1,0), warp_matrix(1,1));
+		cv::Matx22f warp_matrix_2d(warp_matrix(0,0), warp_matrix(0,1), warp_matrix(1,0), warp_matrix(1,1));
 		
 		destination_landmarks = cv::Mat(detected_landmarks.reshape(1, 2).t()) * cv::Mat(warp_matrix_2d).t();
 
@@ -200,19 +201,19 @@ namespace FaceAnalysis
 		destination_landmarks.col(1) = destination_landmarks.col(1) + warp_matrix(1,2);
 		
 		// Move the eyebrows up to include more of upper face
-		destination_landmarks.at<double>(0,1) -= (30/0.7)*sim_scale; 
-		destination_landmarks.at<double>(16,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(0,1) -= (30/0.7)*sim_scale;
+		destination_landmarks.at<float>(16,1) -= (30 / 0.7)*sim_scale;
 
-		destination_landmarks.at<double>(17,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(18,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(19,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(20,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(21,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(22,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(23,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(24,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(25,1) -= (30 / 0.7)*sim_scale;
-		destination_landmarks.at<double>(26,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(17,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(18,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(19,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(20,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(21,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(22,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(23,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(24,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(25,1) -= (30 / 0.7)*sim_scale;
+		destination_landmarks.at<float>(26,1) -= (30 / 0.7)*sim_scale;
 
 		destination_landmarks = cv::Mat(destination_landmarks.t()).reshape(1, 1).t();
 
@@ -363,16 +364,16 @@ namespace FaceAnalysis
 	//===========================================================================
 	// Using Kabsch's algorithm for aligning shapes
 	//This assumes that align_from and align_to are already mean normalised
-	cv::Matx22d AlignShapesKabsch2D(const cv::Mat_<double>& align_from, const cv::Mat_<double>& align_to)
+	cv::Matx22f AlignShapesKabsch2D(const cv::Mat_<float>& align_from, const cv::Mat_<float>& align_to)
 	{
 
 		cv::SVD svd(align_from.t() * align_to);
 
 		// make sure no reflection is there
 		// corr ensures that we do only rotaitons and not reflections
-		double d = cv::determinant(svd.vt.t() * svd.u.t());
+		float d = cv::determinant(svd.vt.t() * svd.u.t());
 
-		cv::Matx22d corr = cv::Matx22d::eye();
+		cv::Matx22f corr = cv::Matx22f::eye();
 		if (d > 0)
 		{
 			corr(1, 1) = 1;
@@ -382,7 +383,7 @@ namespace FaceAnalysis
 			corr(1, 1) = -1;
 		}
 
-		cv::Matx22d R;
+		cv::Matx22f R;
 		cv::Mat(svd.vt.t()*cv::Mat(corr)*svd.u.t()).copyTo(R);
 
 		return R;
@@ -390,22 +391,22 @@ namespace FaceAnalysis
 
 	//=============================================================================
 	// Basically Kabsch's algorithm but also allows the collection of points to be different in scale from each other
-	cv::Matx22d AlignShapesWithScale(cv::Mat_<double>& src, cv::Mat_<double> dst)
+	cv::Matx22f AlignShapesWithScale(cv::Mat_<float>& src, cv::Mat_<float> dst)
 	{
 		int n = src.rows;
 
 		// First we mean normalise both src and dst
-		double mean_src_x = cv::mean(src.col(0))[0];
-		double mean_src_y = cv::mean(src.col(1))[0];
+		float mean_src_x = cv::mean(src.col(0))[0];
+		float mean_src_y = cv::mean(src.col(1))[0];
 
-		double mean_dst_x = cv::mean(dst.col(0))[0];
-		double mean_dst_y = cv::mean(dst.col(1))[0];
+		float mean_dst_x = cv::mean(dst.col(0))[0];
+		float mean_dst_y = cv::mean(dst.col(1))[0];
 
-		cv::Mat_<double> src_mean_normed = src.clone();
+		cv::Mat_<float> src_mean_normed = src.clone();
 		src_mean_normed.col(0) = src_mean_normed.col(0) - mean_src_x;
 		src_mean_normed.col(1) = src_mean_normed.col(1) - mean_src_y;
 
-		cv::Mat_<double> dst_mean_normed = dst.clone();
+		cv::Mat_<float> dst_mean_normed = dst.clone();
 		dst_mean_normed.col(0) = dst_mean_normed.col(0) - mean_dst_x;
 		dst_mean_normed.col(1) = dst_mean_normed.col(1) - mean_dst_y;
 
@@ -416,25 +417,25 @@ namespace FaceAnalysis
 		cv::Mat dst_sq;
 		cv::pow(dst_mean_normed, 2, dst_sq);
 
-		double s_src = sqrt(cv::sum(src_sq)[0] / n);
-		double s_dst = sqrt(cv::sum(dst_sq)[0] / n);
+		float s_src = sqrt(cv::sum(src_sq)[0] / n);
+		float s_dst = sqrt(cv::sum(dst_sq)[0] / n);
 
 		src_mean_normed = src_mean_normed / s_src;
 		dst_mean_normed = dst_mean_normed / s_dst;
 
-		double s = s_dst / s_src;
+		float s = s_dst / s_src;
 
 		// Get the rotation
-		cv::Matx22d R = AlignShapesKabsch2D(src_mean_normed, dst_mean_normed);
+		cv::Matx22f R = AlignShapesKabsch2D(src_mean_normed, dst_mean_normed);
 
-		cv::Matx22d	A;
+		cv::Matx22f	A;
 		cv::Mat(s * R).copyTo(A);
 
-		cv::Mat_<double> aligned = (cv::Mat(cv::Mat(A) * src.t())).t();
-		cv::Mat_<double> offset = dst - aligned;
+		cv::Mat_<float> aligned = (cv::Mat(cv::Mat(A) * src.t())).t();
+		cv::Mat_<float> offset = dst - aligned;
 
-		double t_x = cv::mean(offset.col(0))[0];
-		double t_y = cv::mean(offset.col(1))[0];
+		float t_x = cv::mean(offset.col(0))[0];
+		float t_y = cv::mean(offset.col(1))[0];
 
 		return A;
 
@@ -444,17 +445,17 @@ namespace FaceAnalysis
 	//===========================================================================
 	// Visualisation functions
 	//===========================================================================
-	void Project(cv::Mat_<double>& dest, const cv::Mat_<double>& mesh, double fx, double fy, double cx, double cy)
+	void Project(cv::Mat_<float>& dest, const cv::Mat_<float>& mesh, float fx, float fy, float cx, float cy)
 	{
-		dest = cv::Mat_<double>(mesh.rows, 2, 0.0);
+		dest = cv::Mat_<float>(mesh.rows, 2, 0.0);
 
 		int num_points = mesh.rows;
 
-		double X, Y, Z;
+		float X, Y, Z;
 
 
-		cv::Mat_<double>::const_iterator mData = mesh.begin();
-		cv::Mat_<double>::iterator projected = dest.begin();
+		cv::Mat_<float>::const_iterator mData = mesh.begin();
+		cv::Mat_<float>::iterator projected = dest.begin();
 
 		for (int i = 0; i < num_points; i++)
 		{
@@ -463,8 +464,8 @@ namespace FaceAnalysis
 			Y = *(mData++);
 			Z = *(mData++);
 
-			double x;
-			double y;
+			float x;
+			float y;
 
 			// if depth is 0 the projection is different
 			if (Z != 0)
@@ -484,22 +485,23 @@ namespace FaceAnalysis
 		}
 
 	}
+
 	//===========================================================================
 	// Angle representation conversion helpers
 	//===========================================================================
 
 	// Using the XYZ convention R = Rx * Ry * Rz, left-handed positive sign
-	cv::Matx33d Euler2RotationMatrix(const cv::Vec3d& eulerAngles)
+	cv::Matx33f Euler2RotationMatrix(const cv::Vec3f& eulerAngles)
 	{
-		cv::Matx33d rotation_matrix;
+		cv::Matx33f rotation_matrix;
 
-		double s1 = sin(eulerAngles[0]);
-		double s2 = sin(eulerAngles[1]);
-		double s3 = sin(eulerAngles[2]);
+		float s1 = sin(eulerAngles[0]);
+		float s2 = sin(eulerAngles[1]);
+		float s3 = sin(eulerAngles[2]);
 
-		double c1 = cos(eulerAngles[0]);
-		double c2 = cos(eulerAngles[1]);
-		double c3 = cos(eulerAngles[2]);
+		float c1 = cos(eulerAngles[0]);
+		float c2 = cos(eulerAngles[1]);
+		float c3 = cos(eulerAngles[2]);
 
 		rotation_matrix(0, 0) = c2 * c3;
 		rotation_matrix(0, 1) = -c2 *s3;
@@ -513,53 +515,52 @@ namespace FaceAnalysis
 
 		return rotation_matrix;
 	}
-
+	
 	// Using the XYZ convention R = Rx * Ry * Rz, left-handed positive sign
-	cv::Vec3d RotationMatrix2Euler(const cv::Matx33d& rotation_matrix)
+	cv::Vec3f RotationMatrix2Euler(const cv::Matx33f& rotation_matrix)
 	{
-		double q0 = sqrt(1 + rotation_matrix(0, 0) + rotation_matrix(1, 1) + rotation_matrix(2, 2)) / 2.0;
-		double q1 = (rotation_matrix(2, 1) - rotation_matrix(1, 2)) / (4.0*q0);
-		double q2 = (rotation_matrix(0, 2) - rotation_matrix(2, 0)) / (4.0*q0);
-		double q3 = (rotation_matrix(1, 0) - rotation_matrix(0, 1)) / (4.0*q0);
+		float q0 = sqrt(1 + rotation_matrix(0, 0) + rotation_matrix(1, 1) + rotation_matrix(2, 2)) / 2.0f;
+		float q1 = (rotation_matrix(2, 1) - rotation_matrix(1, 2)) / (4.0f*q0);
+		float q2 = (rotation_matrix(0, 2) - rotation_matrix(2, 0)) / (4.0f*q0);
+		float q3 = (rotation_matrix(1, 0) - rotation_matrix(0, 1)) / (4.0f*q0);
 
-		double t1 = 2.0 * (q0*q2 + q1*q3);
+		float t1 = 2.0f * (q0*q2 + q1*q3);
 
-		double yaw = asin(2.0 * (q0*q2 + q1*q3));
-		double pitch = atan2(2.0 * (q0*q1 - q2*q3), q0*q0 - q1*q1 - q2*q2 + q3*q3);
-		double roll = atan2(2.0 * (q0*q3 - q1*q2), q0*q0 + q1*q1 - q2*q2 - q3*q3);
+		float yaw = asin(2.0 * (q0*q2 + q1*q3));
+		float pitch = atan2(2.0 * (q0*q1 - q2*q3), q0*q0 - q1*q1 - q2*q2 + q3*q3);
+		float roll = atan2(2.0 * (q0*q3 - q1*q2), q0*q0 + q1*q1 - q2*q2 - q3*q3);
 
-		return cv::Vec3d(pitch, yaw, roll);
+		return cv::Vec3f(pitch, yaw, roll);
 	}
 
-	cv::Vec3d Euler2AxisAngle(const cv::Vec3d& euler)
+	cv::Vec3f Euler2AxisAngle(const cv::Vec3f& euler)
 	{
-		cv::Matx33d rotMatrix = Euler2RotationMatrix(euler);
-		cv::Vec3d axis_angle;
+		cv::Matx33f rotMatrix = Euler2RotationMatrix(euler);
+		cv::Vec3f axis_angle;
 		cv::Rodrigues(rotMatrix, axis_angle);
 		return axis_angle;
 	}
 
-	cv::Vec3d AxisAngle2Euler(const cv::Vec3d& axis_angle)
+	cv::Vec3f AxisAngle2Euler(const cv::Vec3f& axis_angle)
 	{
-		cv::Matx33d rotation_matrix;
+		cv::Matx33f rotation_matrix;
 		cv::Rodrigues(axis_angle, rotation_matrix);
 		return RotationMatrix2Euler(rotation_matrix);
 	}
 
-	cv::Matx33d AxisAngle2RotationMatrix(const cv::Vec3d& axis_angle)
+	cv::Matx33f AxisAngle2RotationMatrix(const cv::Vec3f& axis_angle)
 	{
-		cv::Matx33d rotation_matrix;
+		cv::Matx33f rotation_matrix;
 		cv::Rodrigues(axis_angle, rotation_matrix);
 		return rotation_matrix;
 	}
 
-	cv::Vec3d RotationMatrix2AxisAngle(const cv::Matx33d& rotation_matrix)
+	cv::Vec3f RotationMatrix2AxisAngle(const cv::Matx33f& rotation_matrix)
 	{
-		cv::Vec3d axis_angle;
+		cv::Vec3f axis_angle;
 		cv::Rodrigues(rotation_matrix, axis_angle);
 		return axis_angle;
 	}
-
 
 	//============================================================================
 	// Matrix reading functionality
